Blocked revisions 145226 via svnmerge
........
r145226 | russell | 2008-09-30 16:32:53 -0500 (Tue, 30 Sep 2008) | 16 lines

Add support for call pickup on Snom phones.  Asterisk now includes a magic
call-id in the dialog-info event package used with extension state subscriptions
on Snom phones.  Then, when the phone sends an INVITE with Replaces for the
special callid, Asterisk will perform a pickup on the extension that was
subscribed to.

The original code on this issue was submitted by xylome.  However, contributions
have been made by (at least) mgernoth and pkempgen.  The final patch was written
by seanbright, and includes the necessary logic to allow this work in a
technology independent way.

(closes issue #5014)
Reported by: xylome
Patches:
      issue5014-trunk.diff uploaded by seanbright (license 71)

........