Yes, it is true, bkruse is hacking at appdocsxml again!
Things got kind of busy, so I had to stop, but I am converting
and going to help the cause again, besides just the spark at the
beggining of the project. Great job eliel and mvanbaak, for taking
to the next step, a few "next steps", even.

Converted app_mp3 to XML