Merged revisions 146640,146644 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
................
r146640 | mmichelson | 2008-10-06 10:29:56 -0500 (Mon, 06 Oct 2008) | 15 lines

This commit introduces a change to how the "joinempty"
and "leavewhenempty" options are configured in queues.conf.

Instead of using vague terms like "yes," "no," "loose," and
"strict," we now accept a comma-separated list of values
to determine when to consider a member available.

Extended details can be found in the queues.conf.sample
file. Note also that the above four referenced values are
still accepted for backwards-compatibility, but are mapped
internally to the new method of representing the option.

AST-105


................
r146644 | kpfleming | 2008-10-06 11:03:37 -0500 (Mon, 06 Oct 2008) | 15 lines

Merged revisions 146643 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
  r146643 | kpfleming | 2008-10-06 10:57:49 -0500 (Mon, 06 Oct 2008) | 8 lines
  
  ensure that the private structure for pseudo channels is created without 'leaking' configuration data from other configured channels
  
  (closes issue #13555)
  Reported by: jeffg
  Patches:
        issue_13555.patch uploaded by kpfleming (license 421)
  Tested by: jeffg
........

................