I collected, over the last several weeks, several different threadpool implementations in several languages, and gave them a look-at.
And in the end, concluded that the threadpool implementation
in chan_iax2 was head and shoulders above all the others
for the purpose it serves in asterisk.

Most of the other implementations spent a good deal
of energy in providing a queue for jobs awaiting
threads. Submitting a thread involves (usually)
tossing into the queue, and the thread processor
takes any waiting jobs and assigns them to threads
independently.

Well, in Asterisk, this is silly. You can't wait
around for a thread. If there aren't enough threads,
you create a new DYNAMIC thread and get that job
going NOW. DYNAMIC threads are allowed to die if they
sit around unused too long. Otherwise, they get reused.

This DYNAMIC feature is the distinguishing mark of
the IAX2 threadpool implementation. 

In a typical fit of obsessive-compulsive behavior,
I nipped and tucked, I snipped and cut and
pasted, changed names to protect the innocent,
and made a fairly generic implementation out of the IAX2
threadpool code. You create them, and get back a 
pointer. You use that pointer to submit jobs.
You can several different threadpools. You can
set the func pointer in the pool itself, and 
make your threads homogeneous, or you can supply
a func to run in the run method, and make your
pool non-homogeneous. There's a "core show threadpools"
command that shows all your pools and threads (just like
IAX's "iax2 show threadpool", with only a slight diff.

All that's left to do is make pbx_start use a threadpool
to start its stuff, and eval the effect on chan_sip...
and see if the experiment is worth the effort!