Blocked revisions 150307 via svnmerge
........
r150307 | mmichelson | 2008-10-16 19:13:35 -0500 (Thu, 16 Oct 2008) | 14 lines

After a long discussion on #asterisk-bugs, it seems kind of
odd that a channel would be named after the originating port. 
For endpoints that always include ":5060" as part
of the From: header, it will mean that you have a ton of
channels with names like "SIP/5060-3ea38a8b."

I am boldly moving forward with this change in trunk, but I'm
not touching other branches with this one since this definitely
would qualify as a behavior change. If there is a problem with
this commit, and I haven't seen the obvious reason why you'd want
to name the channel after the port from which the call originated,
then please feel free to revert this


........