Merged revisions 150817 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r150817 | bweschke | 2008-10-17 21:18:33 -0500 (Fri, 17 Oct 2008) | 8 lines

 Using the GetVar handler in AMI is potentially dangerous (insta-crash [tm]) when you use a dialplan function that requires a channel and then you don't provide one or provide an invalid one in the Channel: parameter. We'll handle this situation exactly the same way it was handled in pbx.c back on r61766.
We'll create a bogus channel for the function call and destroy it when we're done. If we have trouble allocating the bogus channel then we're not going to try executing the function call at all and run the risk of crashing.

(closes issue #13715)
reported by: makoto
patch by: bweschke


........