These tweaks allow threadpool to work, but...
At about 100 calls/sec, I get crashes. I can process
thousands of calls at slower rates with no problem,
but at this level, I get crashes.

So, I guess, when I get time, I'll check to see
what trunk's max call/sec rate is, to see if the
problem is in the threadpool, or in chan_sip.

It appears to be mem corruption, so I suspect
the locking is at issue.