Merged revisions 151763 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r151763 | twilson | 2008-10-23 11:04:42 -0500 (Thu, 23 Oct 2008) | 9 lines

Backport fix from 1.6.0 that allows you to set parkedcalltransfers=no|caller|callee|both, but default to both which would be the equivalent of the existing behaviour.

The problem was that if someone parked a call, the callee and caller would both get assigned the builtin transfer feature, which would not only be potentially giving someone the ability to transfer themselves when they shouldn't have it, but would also dissallow reinviting the media off of the call.
(closes issue #12854)
	Reported by: davidw
	Patches: 
	      parkingfix4.diff.txt uploaded by otherwiseguy
		  Tested by: davidw, otherwiseguy

........