Merged revisions 155011 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r155011 | mmichelson | 2008-11-06 13:45:52 -0600 (Thu, 06 Nov 2008) | 8 lines

The documentation listed the ability to set 'maxmsg' per
context. The truth is that you can only set this in the general section
or per mailbox. Thus I am updating the sample config file to be more
accurate.

Thanks to sasargen on IRC for bringing up this issue.


........