Merged revisions 156874,156883 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r156874 | mattf | 2008-11-14 10:34:33 -0600 (Fri, 14 Nov 2008) | 1 line

Remove some useless locking and make sure we hangup channels on a link when we get a GRS.
........
r156883 | mmichelson | 2008-11-14 10:53:38 -0600 (Fri, 14 Nov 2008) | 24 lines

Fix some refcounting in app_queue.c and change the
hashing used by app_queue.c to be case-insensitive.
This is accomplished by adding a new case-insensitive
hashing function.

This was necessary to prevent bad refcount errors
(and potential crashes) which would occur due to the
fact that queues were initially read from the config
file in a case-sensitive manner. Then, when a user
issued a CLI command or manager action, we allowed
for case-insensitive input and used that input to 
directly try to find the queue in the hash table. The result
was either that we could not find a queue that was input or
worse, we would end up hashing to a completely bogus value
based on the input.

This commit resolves the problem presented in
issue #13703. However, that issue was reported against
1.6.0. Since this fix introduces a behavior change, I am
electing to not place this same fix in to the 1.6.0 or 1.6.1
branches, and instead will opt for a change which does not
change behavior.


........