Merged revisions 158606 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r158606 | murf | 2008-11-21 16:40:46 -0700 (Fri, 21 Nov 2008) | 19 lines

Merged revisions 158603 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r158603 | murf | 2008-11-21 16:14:50 -0700 (Fri, 21 Nov 2008) | 11 lines

In reference to the fix made for 13871, I was
merging the fix into 1.6.0 and realized I missed
the code in the h-exten block, and didn't catch it
because my test case had the h-exten commented out.

So, this corrects the code I missed, as a 
preventative against another crash report.
Tested with the h-exten defined, all is well.



........

................