Merged revisions 161115 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r161115 | dhubbard | 2008-12-04 17:00:30 -0600 (Thu, 04 Dec 2008) | 8 lines

If 'faxdetect=yes' in sip.conf, switch to a 'fax' extension (if it exists) after T38 is negotiated.  
Terry Wilson created the original patch for this functionality, which I slightly modified and added 
the faxdetect=yes|no configuration option.  This patch is only for T38 fax detection and does not 
do anything for G711 over SIP fax detection.  By default, this option is disabled. 

This functionality is for issue AST-140.


........